requirements.txt
streamlit
pandas
paho-mqtt
scikit-learn
joblib
streamlit-echarts